fix: hubstaff integration redirect (#6802)
