+++ packages/core/src/integration/hubstaff/hubstaff.controller.ts
@@ -10,9 +10,9 @@ import {
 	IOrganization
 } from '@gauzy/contracts';
 import { ApiTags } from '@nestjs/swagger';
-import { Permissions } from './../shared/decorators';
-import { PermissionGuard, TenantPermissionGuard } from './../shared/guards';
-import { UUIDValidationPipe } from './../shared/pipes';
+import { Permissions } from 'shared/decorators';
+import { PermissionGuard, TenantPermissionGuard } from 'shared/guards';
+import { UUIDValidationPipe } from 'shared/pipes';
 import { HubstaffService } from './hubstaff.service';
 
 @ApiTags('Hubstaff Integrations')
