+++ packages/core/src/integration/integration.module.ts
@@ -2,14 +2,15 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { RouterModule } from 'nest-router';
 import { CqrsModule } from '@nestjs/cqrs';
+import { TenantModule } from 'tenant/tenant.module';
+import { UserModule } from 'user/user.module';
+import { HubstaffModule } from './hubstaff/hubstaff.module';
 import { IntegrationType } from './integration-type.entity';
 import { Integration } from './integration.entity';
 import { IntegrationService } from './integration.service';
 import { IntegrationController } from './integration.controller';
 import { CommandHandlers } from './commands/handlers';
 import { IntegrationTenantModule } from '../integration-tenant/integration-tenant.module';
-import { TenantModule } from './../tenant/tenant.module';
-import { UserModule } from './../user/user.module';
 
 @Module({
 	imports: [
@@ -17,7 +18,8 @@ import { UserModule } from './../user/user.module';
 			{
 				path: '/integration', module: IntegrationModule,
 				children: [
-					{ path: '/', module: IntegrationModule },
+					{ path: '/hubstaff', module: HubstaffModule },
+					{ path: '/', module: IntegrationModule }
 				]
 			},
 		]),
@@ -29,6 +31,7 @@ import { UserModule } from './../user/user.module';
 		IntegrationTenantModule,
 		TenantModule,
 		UserModule,
+		HubstaffModule,
 		CqrsModule
 	],
 	controllers: [
